imported/w3c/web-platform-tests/IndexedDB/idb-explicit-commit.any.html is failing https://bugs.webkit.org/show_bug.cgi?id=277659 rdar://133254563
Reviewed by Chris Dumez.

To fix test failure, this patch makes following changes:
1. Add support for simultaneous transactions in SQLiteIDBBackingStore, by only creating a SQLite transaction when the
mode of IDBTransaction is readwrite. With this change, the backend can support simultaneous readonly transactions per
spec. Also, update existing checks in operations that expect SQLite transaction to exists, to expect either SQLite
transaction exists or the transaction is read-only.
2. When IDBTransaction.commit() is invoked during event dispatch for a request, consider the request as handled by
incrementing handledRequestResultsCount. Otherwise, the IndexedDB backend will take the request as not handled by
client, and if the event is error, the backend would abort the transaction instead of committing it. According to spec,
if error event is handled, the transaction can commit; if error event is not handled, the transaction should abort.

This patch also updates some tests that start simultaneous read-only transactions and expect the transactions to happen
in order to start transaction in order (i.e. starting a new transaction when the previous one has finished), to make
the result deterministic.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idb-explicit-commit.any-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idb-explicit-commit.any.js:
* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idb-explicit-commit.any.worker-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/structured-clone.any.worker_101-last-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/IndexedDB/structured-clone.any_101-last-expected.txt:
* LayoutTests/inspector/indexeddb/clearObjectStore.html:
* LayoutTests/inspector/indexeddb/requestData.html:
* LayoutTests/platform/mac/TestExpectations:
* Source/WebCore/Modules/indexeddb/IDBRequest.h:
(WebCore::IDBRequest::isEventBeingDispatched const):
* Source/WebCore/Modules/indexeddb/IDBTransaction.cpp:
(WebCore::IDBTransaction::commitInternal):
* Source/WebCore/Modules/indexeddb/server/IDBBackingStore.h:
* Source/WebCore/Modules/indexeddb/server/MemoryIDBBackingStore.h:
* Source/WebCore/Modules/indexeddb/server/SQLiteIDBBackingStore.cpp:
(WebCore::IDBServer::SQLiteIDBBackingStore::createObjectStore):
(WebCore::IDBServer::SQLiteIDBBackingStore::getRecord):
(WebCore::IDBServer::SQLiteIDBBackingStore::getAllObjectStoreRecords):
(WebCore::IDBServer::SQLiteIDBBackingStore::getAllIndexRecords):
(WebCore::IDBServer::SQLiteIDBBackingStore::getIndexRecord):
(WebCore::IDBServer::SQLiteIDBBackingStore::getCount):
(WebCore::IDBServer::SQLiteIDBBackingStore::openCursor):
(WebCore::IDBServer::SQLiteIDBBackingStore::iterateCursor):
* Source/WebCore/Modules/indexeddb/server/SQLiteIDBBackingStore.h:
* Source/WebCore/Modules/indexeddb/server/SQLiteIDBCursor.cpp:
(WebCore::IDBServer::SQLiteIDBCursor::createSQLiteStatement):
(WebCore::IDBServer::SQLiteIDBCursor::resetAndRebindStatement):
(WebCore::IDBServer::SQLiteIDBCursor::resetAndRebindPreIndexStatementIfNecessary):
(WebCore::IDBServer::SQLiteIDBCursor::internalFetchNextRecord):
(WebCore::IDBServer::SQLiteIDBCursor::iterate):
* Source/WebCore/Modules/indexeddb/server/SQLiteIDBTransaction.cpp:
(WebCore::IDBServer::SQLiteIDBTransaction::begin):
(WebCore::IDBServer::SQLiteIDBTransaction::commit):
(WebCore::IDBServer::SQLiteIDBTransaction::moveBlobFilesIfNecessary):
(WebCore::IDBServer::SQLiteIDBTransaction::deleteBlobFilesIfNecessary):
(WebCore::IDBServer::SQLiteIDBTransaction::abort):
(WebCore::IDBServer::SQLiteIDBTransaction::maybeOpenBackingStoreCursor):
(WebCore::IDBServer::SQLiteIDBTransaction::maybeOpenCursor):
(WebCore::IDBServer::SQLiteIDBTransaction::notifyCursorsOfChanges):
(WebCore::IDBServer::SQLiteIDBTransaction::inProgressOrReadOnly const):
(WebCore::IDBServer::SQLiteIDBTransaction::addBlobFile):
(WebCore::IDBServer::SQLiteIDBTransaction::addRemovedBlobFile):
(WebCore::IDBServer::SQLiteIDBTransaction::sqliteDatabase const):
* Source/WebCore/Modules/indexeddb/server/SQLiteIDBTransaction.h:
(WebCore::IDBServer::SQLiteIDBTransaction::isReadOnly const):
* Source/WebCore/Modules/indexeddb/server/UniqueIDBDatabase.cpp:
(WebCore::IDBServer::UniqueIDBDatabase::takeNextRunnableTransaction):

Canonical link: https://commits.webkit.org/281933@main

+++ Source/WebCore/Modules/indexeddb/server/SQLiteIDBTransaction.cpp
@@ -51,12 +51,16 @@ SQLiteIDBTransaction::~SQLiteIDBTransaction()
     clearCursors();
 }
 
-
 IDBError SQLiteIDBTransaction::begin(SQLiteDatabase& database)
 {
     ASSERT(!m_sqliteTransaction);
 
-    m_sqliteTransaction = makeUnique<SQLiteTransaction>(database, m_info.mode() == IDBTransactionMode::Readonly);
+    if (isReadOnly()) {
+        m_sqliteDatabase = &database;
+        return IDBError { };
+    }
+
+    m_sqliteTransaction = makeUnique<SQLiteTransaction>(database, true);
     m_sqliteTransaction->begin();
 
     if (m_sqliteTransaction->inProgress())
@@ -68,11 +72,16 @@ IDBError SQLiteIDBTransaction::begin(SQLiteDatabase& database)
 IDBError SQLiteIDBTransaction::commit()
 {
     LOG(IndexedDB, "SQLiteIDBTransaction::commit");
+
+    if (isReadOnly()) {
+        reset();
+        return IDBError { };
+    }
+
     if (!m_sqliteTransaction || !m_sqliteTransaction->inProgress())
         return IDBError { ExceptionCode::UnknownError, "No SQLite transaction in progress to commit"_s };
 
     m_sqliteTransaction->commit();
-
     if (m_sqliteTransaction->inProgress())
         return IDBError { ExceptionCode::UnknownError, "Unable to commit SQLite transaction in database backend"_s };
 
@@ -85,6 +94,8 @@ IDBError SQLiteIDBTransaction::commit()
 
 void SQLiteIDBTransaction::moveBlobFilesIfNecessary()
 {
+    ASSERT(!isReadOnly());
+
     String databaseDirectory = m_backingStore.databaseDirectory();
     for (auto& entry : m_blobTemporaryAndStoredFilenames) {
         if (!FileSystem::hardLinkOrCopyFile(entry.first, FileSystem::pathByAppendingComponent(databaseDirectory, entry.second)))
@@ -98,6 +109,8 @@ void SQLiteIDBTransaction::moveBlobFilesIfNecessary()
 
 void SQLiteIDBTransaction::deleteBlobFilesIfNecessary()
 {
+    ASSERT(!isReadOnly());
+
     if (m_blobRemovedFilenames.isEmpty())
         return;
 
@@ -113,6 +126,11 @@ void SQLiteIDBTransaction::deleteBlobFilesIfNecessary()
 
 IDBError SQLiteIDBTransaction::abort()
 {
+    if (isReadOnly()) {
+        reset();
+        return IDBError { };
+    }
+
     for (auto& entry : m_blobTemporaryAndStoredFilenames)
         FileSystem::deleteFile(entry.first);
 
@@ -139,8 +157,7 @@ void SQLiteIDBTransaction::reset()
 
 std::unique_ptr<SQLiteIDBCursor> SQLiteIDBTransaction::maybeOpenBackingStoreCursor(uint64_t objectStoreID, uint64_t indexID, const IDBKeyRangeData& range)
 {
-    ASSERT(m_sqliteTransaction);
-    ASSERT(m_sqliteTransaction->inProgress());
+    ASSERT(inProgressOrReadOnly());
 
     auto cursor = SQLiteIDBCursor::maybeCreateBackingStoreCursor(*this, objectStoreID, indexID, range);
 
@@ -152,12 +169,10 @@ std::unique_ptr<SQLiteIDBCursor> SQLiteIDBTransaction::maybeOpenBackingStoreCurs
 
 SQLiteIDBCursor* SQLiteIDBTransaction::maybeOpenCursor(const IDBCursorInfo& info)
 {
-    ASSERT(m_sqliteTransaction);
-    if (!m_sqliteTransaction->inProgress())
+    if (m_sqliteTransaction && !m_sqliteTransaction->inProgress())
         return nullptr;
 
     auto addResult = m_cursors.add(info.identifier(), SQLiteIDBCursor::maybeCreate(*this, info));
-
     ASSERT(addResult.isNewEntry);
 
     // It is possible the cursor failed to create and we just stored a null value.
@@ -185,6 +200,8 @@ void SQLiteIDBTransaction::closeCursor(SQLiteIDBCursor& cursor)
 
 void SQLiteIDBTransaction::notifyCursorsOfChanges(int64_t objectStoreID)
 {
+    ASSERT(!isReadOnly());
+
     for (auto& i : m_cursors) {
         if (i.value->objectStoreID() == objectStoreID)
             i.value->objectStoreRecordsChanged();
@@ -209,15 +226,32 @@ bool SQLiteIDBTransaction::inProgress() const
     return m_sqliteTransaction && m_sqliteTransaction->inProgress();
 }
 
+bool SQLiteIDBTransaction::inProgressOrReadOnly() const
+{
+    return isReadOnly() || inProgress();
+}
+
 void SQLiteIDBTransaction::addBlobFile(const String& temporaryPath, const String& storedFilename)
 {
+    ASSERT(!isReadOnly());
+
     m_blobTemporaryAndStoredFilenames.append({ temporaryPath, storedFilename });
 }
 
 void SQLiteIDBTransaction::addRemovedBlobFile(const String& removedFilename)
 {
+    ASSERT(!isReadOnly());
     ASSERT(!m_blobRemovedFilenames.contains(removedFilename));
+
     m_blobRemovedFilenames.add(removedFilename);
+}
+
+SQLiteDatabase* SQLiteIDBTransaction::sqliteDatabase() const
+{
+    if (m_sqliteTransaction)
+        return &m_sqliteTransaction->database();
+
+    return m_sqliteDatabase.get();
 }
